rename agent and contract model filenames
